// src/app/model/createquiz.model.ts
export class createquiz {
  id: any = '';
  quizname: string=';'

  questionId: string = '';
  questionName: string = '';
  option1: string = '';
  option2: string = '';
  option3: string = '';
  option4: string = '';
  correct: boolean = false;
}
export class createquizdemo {
  id: any = '';
  quizname: string = ';';

  questionId: string = '';
  questionName: string = '';
  option1: string = '';
  option2: string = '';
  option3: string = '';
  option4: string = '';
  correct: string[] = [];
}